Changing scale for world coordinates from tiles to game units.
Previously game elements's would be defined by a float whose value would
represent the element's location in the tiled map. This means that when
redering elemtns we would need to change the scales. This made it easier
to handle interactions between an element and the tiled map, but
interations between elements would have more overhead.

Since I am planning on moving from tiled based collision to Box2d, I
would like to keep everything using the same coordinates and units.
With this change, all units are now game units. But I have also
disabled movement.

+++ core/src/com/mygdx/game/SingleAssetManager.java
@@ -175,6 +175,6 @@ public class SingleAssetManager {
 
     public interface TextureAnimationReciever {
         void setAnimations(Animation<TextureRegion> walkUp, Animation<TextureRegion> walkRight, Animation<TextureRegion> walkDown, Animation<TextureRegion> walkLeft);
-        void setTextureSize(float width, float height);
+        void setTextureSize(int width, int height);
     }
 }
